refactor(ui): Rename Curatr file name matches to Stream
