+++ ui/src/containers/Visualisations/TemplateStreamProportionOfSocialInteractions/Editor.js
