+++ ui/src/containers/Visualisations/TemplateStreamProportionOfSocialInteractions/buildModel.js
